scrolling works on background now

+++ Dangerzone/src/turkycat/productions/dangerzone/gameloop/Consts.java
@@ -2,6 +2,6 @@ package turkycat.productions.dangerzone.gameloop;
 
 public class Consts
 {
-	public static final int MILLIS_PER_TIME_UNIT = 50;
+	public static final float MILLIS_PER_TIME_UNIT = 50f;
 	public static final float INITIAL_GAME_SPEED = 3f;
 }
